Update OpenLDAP version used for testing
The docker image currently used for OpenLDAP during tests uses debian 10 and
OpenLDAP 2.4.47. OTP support has been added to OpenLDAP in version 2.5.
However, the current image `osixia/openldap` has not had any release since
2021.

This commit change the image used from `osixia/openldap` to a custom
Dockerfile based on alpine. This Dockerfile installs openldap and all of its
overlays, and creates files/directories for slapd to start. The configuration
of the different OpenLDAP databases are done using a set of bootstrapping
scripts in the `test/ldap_data/bootstrap` directory. This allows to customize
the OpenLDAP config and data by adding, editing or deleting LDIF files
from the git repository.

Currently, the configuration is similar to the one used previously. The data
is identical. The LDIF files are loaded using the `entrypoint` script of the
docker container. To determine how to load the LDIF file, the script will look
for special comments starting with `#type: ` to define if we use slapadd or
slapmodify, and `#db: ` to define if we use config or mdb database. See the
`test/ldap_data/README.md` for more details on how to define those values.