Add base support for OpenLDAP OTP overlay
This adds support for OpenLDAP's OTP overlay, only TOTP support has been
added. OpenLDAP's TOTP is implemented as such: the user appends the
6-digit OTP code to the end of their password to authenticate.

This adds multiple configurations:
 - TOTP.Kind: defines the backend used: `db` or `openldap`.
 - TOTP.OpenLDAPParamsDN: defines the DN of the entry containing the
   OTP overlay configuration. This entry must have the `oathTOTPParams`
   object class.

Frontend: no changes made to the API so no update needed.

Regarding the "verify OTP" feature:
When using OpenLDAP as backend for OTP, the OTP code becomes an
intrinsic part of the authentication to LDAP. Therefore, the value of
the "verify" feature is arguable as it is implicitly tested when binding
to LDAP.
However, removing this feature would require adding a field to the
`Features` structure to let the frontend know about its availability
(the feature would be disabled automatically when using the OpenLDAP OTP
backend).

See: slapo-otp(5)