Add support for revoking OTP with mail
Currently, OTP can only be disabled with password authentication.
However, with OpenLDAP OTP, password authentication requires the user to
know its OTP secret. This makes it impossible for a user who has lost the
OTP secret to disable OTP.

This commit adds support for disabling OTP with a token sent by email.
Since this changes the `UserCreateTOTP` struct, the API for
`/totp/create` and `/totp/revoke` is also changed. This also adds the
`/request/revoke` route. The `Password` field was added to the
`Authentication` struct.

Other API endpoints using the Authentication struct have not been
modified to add support for password, for different reasons:
 - Password reinitialization: does not makes sense.
 - Account unlock: arguable because account is often locked due to
   too many failed password attempts.

Frontend was updated and changed format for URL used in mail, the mail
templates have been changed accordingly.