tests: configure static host port for server
With newer versions of Docker, when a container is restarted, a new host
port is selected. When running tests, `GetPort("8000/tcp")` is only
called once at the beginning, the generated URL becomes invalid when the
server is restarted (e.g. to reload configuration file).

This commit configures a static host port so it doesn't changes on
container restart. The `GetFreePortTCP()` helper is used to find an
available TCP port to use.