add hosts mapping host.docker.internal
The IP 172.17.0.1 is the IP of the host on the default bridge on Docker.
However, this bridge doesn't exists when using Podman.

This commit adds an `/etc/hosts` entry to map `host.docker.internal` to
`host-gateway`. This value was added in Docker v20.04 and corresponds to
the IP of the Docker host. On docker it will be 172.17.0.1 when using
the default bridge, on Podman it will be the IP of the host's network
interface.